requirements.txt
asgiref==3.11.0
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.6.0
django-filter==24.3
sqlparse==0.5.5
typing_extensions==4.15.0
